requirements.txt
h5py==3.1.0
numpy==1.17.2
pandas==1.1.5
Pillow==6.1
PyYAML==6.0
scipy==1.5.4
torch==1.3.0
torchvision==0.4.1a0+d94043a
